requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==0.12.2
Flask-Cache==0.13.1
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pylibmc==1.6.0
requests2==2.16.0
urllib3==1.21.1
Werkzeug==0.15.4
